Beginning of large restructuring of music cog.
Restructures the music cog into a module

Renames cogs to cogmanager
Fixes subcommand help invocation
More to assignment expressions in settings
