+++ cogs/music/music_errors.py
