+++ localization/en_en/commands.yaml
@@ -88,6 +88,9 @@ lyrics:
 scrub:
   aliases:        [scrub]
   description:    'Shows a temporary controller/scrubber.'
+reconnect:
+  aliases:        [rc]
+  description:    'disconnects then reconnects the bot in case of bugged songs'
 
 # Misc
 info:
@@ -183,8 +186,8 @@ settings:
       description:      'Set the vote threshold required to skip a song.'
 
 # Cogs
-cogs:
-  aliases:              [cogs]
+cogmanager:
+  aliases:              [cogmanager]
   description:          'cog management'
   sub_commands:
     load:
