--- localization/en_en/nodemanager.yaml
@@ -1,2 +1,0 @@
-pageindicator: Viewing page {_current}/{_total} | Type "{_prefix}help <command>" for more information
-notcommand: Looks like "{_command}" is not a command or category.
